requirements.txt
# Binance合约账户查询工具核心依赖
requests>=2.31.0
python-dotenv>=1.0.0
cryptography>=41.0.0

# 可选依赖 (用于更好的显示效果)
colorama>=0.4.6
tabulate>=0.9.0

# Optional dependencies
matplotlib>=3.7.2
pandas>=2.0.3

# Development dependencies
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0

# Keep existing dependencies below this line

debugpy==1.6.5
decorator==5.1.1
deep-translator==1.11.4
Deprecated==1.2.14
dify-client==0.1.10
dill==0.3.7
discord==2.3.2
discord-oauth2.py==1.2.1
discord.py==2.3.2
distlib==0.3.7
distro==1.8.0
dj-database-url==0.5.0
dnspython==2.3.0
docopt==0.6.2
duckdb==0.8.1
duckduckgo-search==3.8.1
dulwich==0.20.50
email-validator==2.0.0.post2
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.0
execnet==1.9.0
executing==1.2.0
filelock==3.13.0
flake8==6.0.0
flatbuffers==23.5.26
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.6.0
gast==0.5.4
gensim==3.8.0
ghp-import==2.1.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-translate==3.12.1
google-pasta==0.2.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.59.0
grpcio-status==1.59.0
gTTS==2.3.1
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
h5py==3.10.0
hexbytes==0.3.1
hnswlib==0.7.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.5.0
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.0.1
identify==2.5.24
idna==3.3
imageio==2.34.0
importlib_metadata==7.1.0
iniconfig==2.0.0
ipykernel==6.20.1
ipython==8.8.0
isort==5.12.0
itsdangerous==2.1.2
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.18.2
Jinja2==3.1.2
jiter==0.8.2
jmespath==1.0.1
joblib==1.2.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==5.1.3
jwt==1.3.1
kafka-python==2.0.2
keras==3.1.1
keyring==25.1.0
kiwisolver==1.4.4
lanarky==0.7.8
langchain==0.0.206
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.6
langchainplus-sdk==0.0.16
langcodes==3.3.0
langdetect==1.0.9
langsmith==0.0.87
langup==0.0.1
libclang==18.1.1
lockfile==0.12.2
loguru==0.7.0
lru-dict==1.2.0
lxml==4.9.2
lz4==4.3.2
Markdown==3.3.7
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mergedeep==1.3.4
mkdocs==1.4.3
ml-dtypes==0.3.2
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msg-parser==1.2.0
msgpack==1.0.8
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.9
namex==0.0.7
nest-asyncio==1.5.6
nltk==3.6.5
nodeenv==1.8.0
numexpr==2.8.4
numpy==1.25.2
oauthlib==3.2.2
olefile==0.46
onnxruntime==1.15.1
opt-einsum==3.3.0
optree==0.11.0
orjson==3.8.10
outcome==1.2.0
overrides==7.3.1
packaging==23.2
pandas==2.0.3
parsimonious==0.9.0
parso==0.8.3
passlib==1.7.4
pathspec==0.11.1
pathy==0.10.1
patsy==0.5.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
pinecone-client==2.2.1
pipreqs==0.4.11
pkginfo==1.10.0
platformdirs==2.6.2
playsound==1.2.2
plotly==5.12.0
pluggy==1.0.0
poetry==1.2.0
poetry-core==1.1.0
poetry-plugin-export==1.1.2
posthog==3.0.1
pre-commit==3.3.2
preshed==3.0.8
prompt-toolkit==3.0.36
proto-plus==1.22.3
protobuf==4.25.3
psutil==5.9.5
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pulsar-client==3.2.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.20.0
pycryptodomex==3.17
pydantic==1.10.8
pydub==0.25.1
pyflakes==3.0.1
Pygments==2.14.0
PyJWT==2.8.0
pylev==1.4.0
pymdown-extensions==10.0.1
pyOpenSSL==23.2.0
pypandoc==1.11
pyparsing==3.0.6
pypdf==3.11.1
pypdfium2==4.18.0
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-benchmark==4.0.0
pytest-cache==1.0
pytest-cov==4.1.0
pytest-integration==0.2.3
pytest-mock==3.10.0
pytest-recording==0.12.2
pytest-xdist==3.3.1
python-dateutil==2.8.2
python-docx==0.8.11
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
python-slugify==8.0.1
python-telegram-bot==21.10
pyttsx==1.1
pyttsx3==2.71
pytube==15.0.0
pytz==2021.3
pyunormalize==15.1.0
PyYAML==6.0
pyyaml_env_tag==0.1
pyzmq==24.0.1
readability-lxml==0.8.1
redis==4.5.5
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==13.0.1
rlp==4.0.0
rsa==4.9
s3transfer==0.6.0
schedule==1.2.0
scikit-learn==1.3.0
scikit-posthocs==0.7.0
scipy==1.11.1
seaborn==0.12.2
secure-smtplib==0.1.1
selenium==4.8.2
shellingham==1.5.0.post1
simplejson==3.19.1
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
sniffio==1.3.0
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.27
sqlalchemy2-stubs==0.0.2a35
sqlalchemy_database==0.1.0
sqlmodel==0.0.8
sqlmodelx==0.0.5
sqlparse==0.4.3
srsly==2.4.6
stack-data==0.6.2
starlette==0.27.0
statsmodels==0.13.5
sympy==1.12
ta==0.11.0
TA-Lib==0.4.32
tabulate==0.9.0
tenacity==8.1.0
termcolor==2.4.0
text-to-speech==1.6.0
text-unidecode==1.3
thinc==8.2.3
threadpoolctl==3.0.0
tiktoken==0.6.0
tokenizers==0.13.2
tomlkit==0.12.4
toolz==0.12.1
tornado==6.2
tqdm==4.66.2
traitlets==5.8.1
transformers==4.25.1
trio==0.22.0
trio-websocket==0.9.2
trueskill==0.4.5
tushare==1.2.89
tweepy==4.14.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.0.1
unstructured==0.7.3
uritemplate==4.1.1
urllib3==1.26.7
urllib3-secure-extra==0.1.0
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
vcrpy==4.3.1
virtualenv==20.21.1
wasabi==1.1.1
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.5
weasel==0.3.4
weaviate-client==3.19.2
web3==6.15.1
webdriver-manager==3.8.6
webencodings==0.5.1
websocket-client==0.57.0
websockets==11.0.3
Werkzeug==3.0.1
wikipedia==1.4.0
wrapt==1.14.1
wsproto==1.2.0
xattr==0.9.9
xgboost==2.1.0
xlrd==2.0.1
XlsxWriter==3.1.2
xmltodict==0.13.0
zipp==3.6.0
zstandard==0.21.0
